backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
httpx==0.26.0
sqlalchemy==2.0.25
alembic==1.13.1
passlib==1.7.4
python-multipart==0.0.7
psycopg2-binary==2.9.9
authlib==1.3.0
asyncpg==0.29.0
itsdangerous==2.1.2
bcrypt==4.1.2
pytest==7.4.3
python-dotenv==1.0.0
supabase==2.3.3 